+++ lib/models/base_bridge_signal.dart
@@ -1,8 +1,8 @@
 import 'package:trinity/models/signal.dart';
 import 'package:trinity/node_anatomy.dart';
 
-abstract class BaseBridgeSignal<V> extends Signal<V?> {
-  BaseBridgeSignal(super.value);
+abstract class BaseBridgeSignal<V> extends Signal<V> {
+  BaseBridgeSignal() : super.deferred();
 
   void connect(InheritedTrinityScope scope);
 }