+++ lib/models/base_signal.dart
@@ -7,12 +7,18 @@ import 'package:trinity/trinity.dart';
 /// Internal base class containing the "raw" state and stream logic.
 /// Signal extends this, but ReadableSignal only wraps it.
 abstract class BaseSignal<T> {
-  T _value;
+  late T _value;
+  bool _isInitialized = false;
   final controller = StreamController<T>.broadcast();
 
   late final Node attachedNode;
 
-  BaseSignal(this._value);
+  BaseSignal(T value) {
+    _value = value;
+    _isInitialized = true;
+  }
+
+  BaseSignal.deferred();
 
   T get value => _value;
 
@@ -20,11 +26,16 @@ abstract class BaseSignal<T> {
   T get unsafeValue => _value;
 
   @protected
-  set unsafeValue(T v) => _value = v;
+  set unsafeValue(T v) {
+    _value = v;
+    _isInitialized = true;
+  }
 
   Stream<T> get stream => controller.stream;
 
-  Stream<T> get streamTriggerImmediatly => controller.stream.startWith(_value);
+  Stream<T> get streamTriggerImmediatly => _isInitialized 
+      ? controller.stream.startWith(_value)
+      : controller.stream;
 
   @mustCallSuper
   void dispose() {